Event detection strategy updated

+++ android/app/src/main/java/xyz/praveen/iring/util/OnHitrateChangeListener.java
@@ -11,5 +11,5 @@ public interface OnHitrateChangeListener {
      *
      * @param hitrate New hitrate
      */
-    void onHitrateUpdate(int hitrate);
+    void onHitrateUpdate(float hitrate);
 }
